+++ lib/activerecord/materialized.rb
@@ -41,6 +41,8 @@ require_relative "materialized/relation_cache_writer"
 require_relative "materialized/maintenance_store"
 require_relative "materialized/incremental_maintainer"
 require_relative "materialized/metadata"
+require_relative "materialized/partition_record"
+require_relative "materialized/partition_state"
 
 require_relative "materialized/refresh_job" if defined?(ActiveJob::Base)